requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-chroma==0.1.0

# PDF processing
PyPDF2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0

# Vector database
chromadb==0.4.22
faiss-cpu==1.7.4

# LLM and embeddings
openai==1.6.1
tiktoken==0.5.2
sentence-transformers==2.2.2

# Web interface
streamlit==1.29.0
gradio==4.8.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.24.3
pandas==2.0.3

# Optional: For advanced PDF processing
pdfplumber==0.10.3
pypdf==3.17.4

# Development
jupyter==1.0.0
ipykernel==6.29.0

